Makes time logs store floats instead of integers. Migrate DB for this to take effect

+++ database/migrations/2017_01_09_160656_create_time_logs_table.php
@@ -17,7 +17,7 @@ class CreateTimeLogsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('Individual_Report_id');
             $table->date('Day');
-            $table->integer('Hours');
+            $table->float('Hours');
             $table->string('Description');
             $table->timestamps();
